+++ mycodo/mycodo_flask/api/__init__.py
@@ -67,6 +67,7 @@ def init_api(app):
     import mycodo.mycodo_flask.api.function
     import mycodo.mycodo_flask.api.input
     import mycodo.mycodo_flask.api.measurement
+    import mycodo.mycodo_flask.api.note
     import mycodo.mycodo_flask.api.output
     import mycodo.mycodo_flask.api.pid
     import mycodo.mycodo_flask.api.settings